[QCF-80] 게시물 상세 이미지위에 좋아요 버튼 추가 번경
