+++ desk/src/components/units/boards/components/detailBoardImages/types.ts
@@ -1,0 +1,7 @@
+import { TUser } from '@/src/commons/types/generated/types'
+
+export type DetailBoardImagesProps = {
+  boardId: string
+  imageURLs: string[]
+  likers: TUser[]
+}
