+++ desk/src/components/units/boards/detail/Detail.presenter.tsx
@@ -7,6 +7,7 @@ import ProductItemCardList from '../components/productItemCardList'
 import ProfileHeader from '../components/profileHeader'
 import TitleWithDescription from '../components/titleWithDescription'
 import { BoardDetailUIProps } from './Detail.types'
+import DetailBoardImages from '../components/detailBoardImages'
 
 export default function BoardDetailUI(props: BoardDetailUIProps) {
   return (
@@ -18,7 +19,12 @@ export default function BoardDetailUI(props: BoardDetailUIProps) {
           userData={props.boardData.writer}
         />
       </Box>
-      <Carousel imageURLs={props.boardData.pictures.map(item => item.url)} />
+      <DetailBoardImages
+        boardId={props.boardData.id}
+        likers={props.boardData.likers ?? []}
+        imageURLs={props.boardData.pictures.map(item => item.url)}
+      />
+      {/* <Carousel imageURLs={props.boardData.pictures.map(item => item.url)} /> */}
       <Box pt={'8px'}>
         <CountStateWithLike
           boardId={props.boardData.id}
